Get rid of arch/mips64/kernel.  9116 lines of code gone.

+++ include/asm-mips/processor.h
@@ -234,7 +234,7 @@ struct task_struct;
 /* Prepare to copy thread state - unlazy all lazy status */
 #define prepare_to_copy(tsk)	do { } while (0)
 
-extern int kernel_thread(int (*fn)(void *), void * arg, unsigned long flags);
+extern long kernel_thread(int (*fn)(void *), void * arg, unsigned long flags);
 
 extern unsigned long thread_saved_pc(struct task_struct *tsk);
